requirements.txt
asgiref==3.7.2
astroid==3.0.1
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.1
dill==0.3.7
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pycodestyle==2.11.1
PyJWT==1.7.1
pylint==3.0.2
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
supervisor==4.2.5
tomlkit==0.12.1
urllib3==2.0.7
